rag-mcp-example/requirements.txt
# Core dependencies for RAG vs MCP examples

# Anthropic API
anthropic>=0.39.0

# RAG Dependencies
pypdf2>=3.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0

# MCP Dependencies
mcp>=1.0.0

# FastAPI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

